Use timeout deadline when stopping ThreadedSessionImpl
ThreadedSessionImpl#joinAndClose waits infinitely on the Future -
this should be avoided. Also both stop and joinAndClose methods
have to be called for proper cleanup - it would be better to combine
them. I assume the purpose of separating them is to allow one to
initiate the stop and then wait for completion later. However this
can be accomplished easier by having stop return a Future. This also
allows the caller to choose the timeout deadline which makes sense
since they're initiating the stop and know if/how they should wait for
completion.

Change-Id: I5a070ea121fda3e90c1f60491594b13541c99fd6
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>