JSONRPC-10: Close ZMQ socket properly
The actual fix to is to let zmqContext destroy the socket,
rather than calling socket.close() directly.

However, during fix I noticed that every open() call
reallocates the RX and TX pollers. So added protection to
verify those are null. And added code to unregister socket
from pollers if it is being closed.

Bug: JSONRPC-10
Change-Id: I290da0e6c66460563e0c1f22075a72c12498d96c
Signed-off-by: Shaleen Saxena <shaleen.external@gmail.com>