Remove deprecated AsyncWriteTransaction#commit method
This method has been deprecated for quite some time and now has a
default implementation so can be safely removed from sub-classes.

I also found that JsonRPCTx.submit did not fail the returned Future
on RemoteOmShard failure (ie commit() returns false). I assume this was
an oversight so I fixed that and wrote a UT for it.

Change-Id: Ia73a5f6badca20e875ba6166e57714cc182fc17f
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>