Eliminate use of controller BindingToNormalizedNodeCodec
EnsureParentTransactionFactory was the only user of the
BindingToNormalizedNodeCodec in the production code - the
code was changed in a manner that doesn't need the
BindingToNormalizedNodeCodec.

UT's were changed to use the mdsal BindingToNormalizedNodeCodec.
In doing so, the DataBrokerTestCustomizer was removed in favor
of the version from the conttoller - not sure why it was copied -
the only difference was the creation of the DOMMountPointService
but that can simply be done outside the DataBrokerTestCustomizer.

Change-Id: Ibf9756acff78c0e4f962e3f1c8f13d0b2a3ed32e
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>