Fix Regression in Json Converter
The fix for root level lists introduced a regression due
to the code for root level being invoked for any list.

The "is this root" check should be done only once and at
root level. From there on the serialization should procede
as usual.

Instead of using a test model the list test now reuses one
of the tests using the topology model. That test was checking
data, it now checks the path as well.

Change-Id: I0c0e8a5d762722869d6abdd093b12072ece37c5e
Signed-off-by: Anton Ivanov <anton.ivanov@cambridgegreys.com>