Improve JSONRPC-DOM codec interactions
JsonConverter class is already convoluted with code that is hard
to read and even harder to maintain. Create JsonRpcPathCodec that can be
used to convert between YangInstanceIdentifier and JSONRPC path.
This means we can remove YangInstanceIdentifierDeserializer too.
Provide JsonRpcPathBuilder for convienient creation of JSONRPC paths.

Remove NormalizedNodesHelper test code as it can be replaced
by simpler means.

Change-Id: Ifcdbf66c7ecf77652e0d459bafccc4d29aed1f0d
Signed-off-by: Richard Kosegi <richard.kosegi@gmail.com>