JSONRPC-41 : Publish operational state even in case of error
Currently there is no way to for user to see cause of failure
of mount operation other then diagnosing log file.

Publish intermediate operational state for peer as well
as failure state

When mount operation fail, entry still needs to be maintained
to allow proper cleanup. This is good as it brings better visibility
into what's going on.

Change-Id: I2dbcf3caca1fe462f7724b922d7112fcbf639bd4
Signed-off-by: Richard Kosegi <richard.kosegi@gmail.com>