Do not fail on warnings for docs-linkcheck
This is a workaround due to sphinx-tabs being incompatible with the
Sphinx linkchecker which throws a warning. We mainly use linkchecker
to catch broken links so we don't really need to fail on warnings as
the main "docs" builder will still fail for warning level issues so
this is a reasonable workaround.

Change-Id: I3cdaf50c59c226430e9dbaf6196a9346ba62e36f
Signed-off-by: Thanh Ha <zxiiro@gmail.com>